--- frontend/techpick/src/queries/index.ts
@@ -1,5 +1,0 @@
-export { useCreateTag } from './useCreateTag';
-export { useDeleteTag } from './useDeleteTag';
-export { useFetchTagList } from './useFetchTagList';
-export { useUpdateTag } from './useUpdateTag';
-export { tagKeys } from './tagKeys';
